+++ lib/pavilion/test_config/__init__.py
@@ -2,12 +2,4 @@ from . import file_format
 from .variables import DeferredVariable, VariableSetManager, VariableError
 from . import variables
 from . import string_parser
-from .setup import (apply_overrides,
-                    load_test_configs,
-                    find_all_tests,
-                    resolve_permutations,
-                    resolve_config,
-                    resolve_deferred,
-                    was_deferred,
-                    resolve_section_vars)
-from .setup import TestConfigError
+from .resolver import TestConfigError, TestConfigResolver
